+++ src/components/ArticleWidget.js
@@ -1,0 +1,86 @@
+import React, { Component } from "react";
+import _ from "lodash";
+import { translate } from "react-i18next";
+import FacebookPlayer from "react-facebook-player";
+import YouTube from "react-youtube";
+import cms from "../content/cms";
+import "./ArticleWidget.css";
+
+const APP_ID = cms.siteConfig.appId;
+const Remarkable = require("remarkable");
+const md = new Remarkable("full", {
+	html: true,
+	linkify: true,
+	typographer: true,
+	breaks: true,
+});
+
+class ArticleWidget extends Component {
+    renderVideo(article) {
+		const { url } = article.fields;
+
+		if (/facebook.com/.test(url)) {
+			let videoId = url.replace(/.*facebook.com\/.*\/videos\/(.*)\/.*/, "$1");
+
+			return <FacebookPlayer className={"Facebook"} videoId={videoId} appId={APP_ID} />;
+		} else if (/youtube.com/) {
+			let videoId = url.replace(/^.*((youtu.be\/)|(v\/)|(\/u\/\w\/)|(embed\/)|(watch\?))\??v?=?([^#&?]*).*/, "$7");
+			return <YouTube videoId={videoId} className={"YouTube"} />;
+		}
+		return null;
+    }
+    
+    render(){        
+
+        const { country, onNavigate, t, article, category, showHero, showFullArticle } = this.props;
+		if (!article) {
+			// Anti pattern, but saves 1 or more ifs.
+			return null;
+		}
+
+		let categorySlug = "article";
+		if (category) {
+			categorySlug = category.fields.slug;
+		}
+
+        let content = showFullArticle ? article.fields.content : article.fields.lead;
+        let showFullArt = showFullArticle;
+		const { contentType } = article.sys;
+		if (contentType.sys.id === "video") {
+			content = article.fields.lead;
+			showFullArt = true;
+		}
+        let hero = article.fields.hero;
+        
+        return (
+            <div className="Article" key={article.sys.id}>
+				{hero &&
+					hero.fields &&
+					hero.fields.file &&
+					showHero && (
+						<div className="hero">
+							<img src={article.fields.hero.fields.file.url + "?fm=jpg&fl=progressive"} alt="" />
+						</div>
+					)}
+				{showFullArt ? <h3>{article.fields.title}</h3> : <h3 onClick={() => onNavigate(`/${country.fields.slug}/${categorySlug}/${article.fields.slug}`)}>{article.fields.title}</h3>}
+				{contentType.sys.id === "video" && this.renderVideo(article)}
+				<p dangerouslySetInnerHTML={{ __html: md.render(content) }} />
+				{!showFullArt && (
+					<s className="Read-More">
+						<a
+							href="#"
+							onClick={() => {
+								onNavigate(`/${country.fields.slug}/${categorySlug}/${article.fields.slug}`);
+								return false;
+							}}
+						>
+							{t("Read More")}
+						</a>
+					</s>
+				)}
+			</div>
+        );
+    }
+}
+
+export default translate()(ArticleWidget);
