+++ src/components/ArticleWidget.scss
@@ -1,0 +1,67 @@
+.Article {
+    .Facebook {
+        width: 100%;
+    }
+
+    .YouTube {
+        width: 100%;
+    }
+    h3 {
+        margin-bottom: 10px;
+    }
+    p {
+        margin-top: 0;
+        font-weight: 300;
+        u {
+            text-decoration: none;
+        }
+    }
+
+    .hero {
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
+
+    .hero > img {
+        max-width: 100%;
+    }
+
+    .pos,
+    .neg {
+        display: none;
+        visibility: hidden;
+        height: 0;
+        width: 0;
+        margin: 0;
+        padding: 0;
+    }
+    .pos + ul,
+    .neg + ul {
+        padding: 0;
+
+        li {
+            list-style: none;
+
+            &::before {
+                font-family: "FontAwesome";
+            }
+        }
+    }
+    .pos + ul {
+        li {
+            &::before {
+                content: "\f00c ";
+                color: green;
+            }
+        }
+    }
+    .neg + ul {
+        li {
+            &::before {
+                content: "\f00d ";
+                color: red;
+            }
+        }
+    }
+}
