+++ src/components/CategoryWidget.scss
@@ -1,0 +1,20 @@
+@import "../shared/includes.scss";
+
+.Category {
+    h3 {
+        text-transform: uppercase;
+    }
+    h3:hover {
+        text-transform: uppercase;
+        text-decoration: underline;
+        cursor: pointer;
+    }
+    .Article {
+        font-size: 80%;
+        h3 {
+            text-transform: none;
+            margin-bottom: 3px;
+            font-weight: 700;
+        }
+    }
+}
