+++ src/components/HomeWidget.js
@@ -3,10 +3,13 @@ import moment from "moment";
 import _ from "lodash";
 
 import { translate } from "react-i18next";
-import FacebookPlayer from "react-facebook-player";
 import YouTube from "react-youtube";
 import cms from "../content/cms";
 import "./HomeWidget.css";
+import {TopCategoriesWidget} from "./";
+import {LocalGuideWidget} from "./";
+import {ArticleWidget} from "./";
+import {CategoryWidget} from "./";
 
 const APP_ID = cms.siteConfig.appId;
 const Remarkable = require("remarkable");
@@ -17,8 +20,7 @@ const md = new Remarkable("full", {
 	breaks: true,
 });
 
-class HomeWidget extends Component {
-	// Maybe these can be a Separate Component?
+class HomeWidget extends Component {	
 
 	renderWidget(w) {
 		if (w.fields.type === "Latest Article of Category") {
@@ -49,45 +51,12 @@ class HomeWidget extends Component {
 		/*jshint ignore:start*/
 		/*eslint-disable*/
 		return (
-			<div className="LocalGuide">
-				<s>
-					<a
-						href="#"
-						onClick={() => {
-							onNavigate(`/${country.fields.slug}/services`);
-							return false;
-						}}
-					>
-						{t("See More")}
-					</a>
-				</s>
-				<h3>{t("Local Guide")}</h3>
-				<div className="Container">
-					{guideItems.map(c => {
-						let image =
-							c.fields.backgroundImage && c.fields.backgroundImage.fields.file ? (
-								<img alt={c.fields.title} src={`${c.fields.backgroundImage.fields.file.url}?fm=jpg&fl=progressive`} />
-							) : (
-								<img alt={c.fields.title} src="https://upload.wikimedia.org/wikipedia/en/4/48/Blank.JPG" />
-							);
-						let link = c => {
-							if (c.fields.url.indexOf("/") === 0) {
-								return onNavigate(c.fields.url);
-							} else {
-								return (global.document.location = c.fields.url);
-							}
-						};
-						return (
-							<div key={c.sys.id} className="LocalGuideItem">
-								{image}
-								<div className="Overlay" onClick={link.bind(null, c)}>
-									{c.fields.title}
-								</div>
-							</div>
-						);
-					})}
-				</div>
-			</div>
+			<LocalGuideWidget
+				country = {country}
+				onNavigate = {onNavigate}
+				t = {t}	
+				guideItems = {guideItems}				
+			/>
 		);
 		/*eslint-enable*/
 		/*jshint ignore:end*/
@@ -99,129 +68,49 @@ class HomeWidget extends Component {
 
 		/*jshint ignore:start*/
 		/*eslint-disable*/
-		return (
-			<div className="TopCategories">
-				<s>
-					<a
-						href="#"
-						onClick={() => {
-							onNavigate(`/${country.fields.slug}/categories`);
-							return false;
-						}}
-					>
-						{t("See More")}
-					</a>
-				</s>
-				<h3>{t("Top Categories")}</h3>
-				{categories.map(c => {
-					let article = articleFunc(c);
-					return (
-						<div key={c.sys.id} className="TopCategory" onClick={() => onNavigate(`/${country.fields.slug}/${c.fields.slug}/${article.fields.slug}`)}>
-							<div className="icon">
-								<i className={c.fields.iconClass || "material-icons"}>{c.fields.iconText || ((!c.fields.iconClass || c.fields.iconClass === "material-icons") && "add")}</i>
-							</div>
-							{c.fields.name}
-						</div>
-					);
-				})}
-			</div>
+		return (			
+			<TopCategoriesWidget
+				country = {country}
+				onNavigate = {onNavigate}	
+				categories = {categories}			
+			/>
 		);
 		/*eslint-enable*/
 		/*jshint ignore:end*/
-	}
-
-	renderVideo(article) {
-		const { url } = article.fields;
-
-		if (/facebook.com/.test(url)) {
-			let videoId = url.replace(/.*facebook.com\/.*\/videos\/(.*)\/.*/, "$1");
-
-			return <FacebookPlayer className={"Facebook"} videoId={videoId} appId={APP_ID} />;
-		} else if (/youtube.com/) {
-			let videoId = url.replace(/^.*((youtu.be\/)|(v\/)|(\/u\/\w\/)|(embed\/)|(watch\?))\??v?=?([^#&?]*).*/, "$7");
-			return <YouTube videoId={videoId} className={"YouTube"} />;
-		}
-		return null;
-	}
+	}	
 
 	renderArticle(article, category, showHero = true, showFullArticle = false) {
-		const { country, onNavigate, t } = this.props;
-		if (!article) {
-			// Anti pattern, but saves 1 or more ifs.
-			return null;
-		}
-
-		let categorySlug = "article";
-		if (category) {
-			categorySlug = category.fields.slug;
-		}
-
-		let content = showFullArticle ? article.fields.content : article.fields.lead;
-		const { contentType } = article.sys;
-		if (contentType.sys.id === "video") {
-			content = article.fields.lead;
-			showFullArticle = true;
-		}
-		let hero = article.fields.hero;
+		const { country, onNavigate, t } = this.props;		
 
 		/*jshint ignore:start*/
 		/*eslint-disable*/
 		return (
-			<div className="Article" key={article.sys.id}>
-				{hero &&
-					hero.fields &&
-					hero.fields.file &&
-					showHero && (
-						<div className="hero">
-							<img src={article.fields.hero.fields.file.url + "?fm=jpg&fl=progressive"} alt="" />
-						</div>
-					)}
-				{showFullArticle ? <h3>{article.fields.title}</h3> : <h3 onClick={() => onNavigate(`/${country.fields.slug}/${categorySlug}/${article.fields.slug}`)}>{article.fields.title}</h3>}
-				{contentType.sys.id === "video" && this.renderVideo(article)}
-				<p dangerouslySetInnerHTML={{ __html: md.render(content) }} />
-				{!showFullArticle && (
-					<s className="Read-More">
-						<a
-							href="#"
-							onClick={() => {
-								onNavigate(`/${country.fields.slug}/${categorySlug}/${article.fields.slug}`);
-								return false;
-							}}
-						>
-							{t("Read More")}
-						</a>
-					</s>
-				)}
-			</div>
+			<ArticleWidget
+				country = {country}
+				onNavigate = {onNavigate}
+				t = {t}
+				article = {article}
+				category = {category}
+				showhero = {showHero}
+				showFullArticle = {showFullArticle}
+			/>
 		);
 		/*eslint-enable*/
 		/*jshint ignore:end*/
 	}
 
 	renderCategory(c) {
-		const { country, onNavigate, t } = this.props;
-
-		let html = md.render(c.fields.description);
-		let article = c.fields.overview || _.first(c.fields.articles);
+		const { country, onNavigate, t } = this.props;	
 
 		/*jshint ignore:start*/
 		/*eslint-disable*/
 		return (
-			<div className="Category">
-				<h3>{c.fields.name}</h3>
-				<p dangerouslySetInnerHTML={{ __html: html }} />
-				<s>
-					<a
-						href="#"
-						onClick={() => {
-							onNavigate(`/${country.fields.slug}/${c.fields.slug}/${article.fields.slug}`);
-							return false;
-						}}
-					>
-						{t("Read More")}
-					</a>
-				</s>
-			</div>
+			<CategoryWidget
+				country = {country}
+				onNavigate = {onNavigate}
+				t = {t}	
+				c = {c}		
+			/>						
 		);
 		/*eslint-enable*/
 		/*jshint ignore:end*/
