+++ src/components/HomeWidget.scss
@@ -27,103 +27,7 @@
 	s.Read-More {
 		display: flex;
 		flex-direction: row;
-	}
-	.TopCategories {
-		h3 {
-			text-transform: uppercase;
-		}
-		.ltr & {
-			s {
-				float: right;
-			}
-		}
-		.rtl & {
-			s {
-				float: left;
-			}
-		}
-	}
-	.LocalGuide {
-		.ltr & {
-			s {
-				float: right;
-			}
-		}
-		.rtl & {
-			s {
-				float: left;
-			}
-		}
-		.Container {
-			display: flex;
-			flex-wrap: wrap;
-			width: 100%;
-
-			.LocalGuideItem {
-				display: block;
-				flex-basis: calc(50% - 5px);
-				margin-bottom: 5px;
-				margin-right: 10px;
-				.rtl & {
-					margin-left: 10px;
-					margin-right: 0;
-				}
-				text-align: center;
-				justify-content: center;
-
-				img {
-					height: 72px;
-					width: 100%;
-					object-fit: cover;
-				}
-				&:nth-child(odd) {
-					color: red;
-					margin-right: 5px;
-					margin-left: 0;
-
-					.rtl & {
-						margin-left: 5px;
-						margin-right: 0;
-					}
-				}
-
-				&:nth-child(even) {
-					color: blue;
-					margin-left: 5px;
-					margin-right: 0;
-
-					.rtl & {
-						margin-right: 5px;
-						margin-left: 0;
-					}
-				}
-			}
-			.Overlay:hover,
-			.Overlay:active {
-				background-color: rgba(0, 0, 0, 0.4);
-			}
-
-			.Overlay {
-				display: flex;
-				position: relative;
-				top: -76px;
-				.rtl & {
-					top: -80px;
-				}
-				background-color: rgba(0, 0, 0, 0.7);
-				height: 72px;
-				justify-content: center;
-				align-content: center;
-				align-items: center;
-				margin-bottom: -76px;
-				color: white;
-				font-size: 14px;
-				letter-spacing: 0;
-				line-height: 16px;
-				transition: background-color 1s ease-in-out;
-			}
-		}
-	}
+	}	
 
 	.TopCategory {
 		display: flex;
@@ -157,92 +61,5 @@
 				padding-left: 10px;
 			}
 		}
-	}
-
-	.Category {
-		h3 {
-			text-transform: uppercase;
-		}
-		h3:hover {
-			text-transform: uppercase;
-			text-decoration: underline;
-			cursor: pointer;
-		}
-		.Article {
-			font-size: 80%;
-			h3 {
-				text-transform: none;
-				margin-bottom: 3px;
-				font-weight: 700;
-			}
-		}
-	}
-
-	.Article {
-		.Facebook {
-			width: 100%;
-		}
-
-		.YouTube {
-			width: 100%;
-		}
-		h3 {
-			margin-bottom: 10px;
-		}
-		p {
-			margin-top: 0;
-			font-weight: 300;
-			u {
-				text-decoration: none;
-			}
-		}
-
-		.hero {
-			display: flex;
-			align-items: center;
-			justify-content: center;
-		}
-
-		.hero > img {
-			max-width: 100%;
-		}
-
-		.pos,
-		.neg {
-			display: none;
-			visibility: hidden;
-			height: 0;
-			width: 0;
-			margin: 0;
-			padding: 0;
-		}
-		.pos + ul,
-		.neg + ul {
-			padding: 0;
-
-			li {
-				list-style: none;
-
-				&::before {
-					font-family: "FontAwesome";
-				}
-			}
-		}
-		.pos + ul {
-			li {
-				&::before {
-					content: "\f00c ";
-					color: green;
-				}
-			}
-		}
-		.neg + ul {
-			li {
-				&::before {
-					content: "\f00d ";
-					color: red;
-				}
-			}
-		}
-	}
+	}	
 }
